requirements.txt
# SYNCTACLES Requirements
# Python 3.12

# Web Framework
fastapi==0.115.7
uvicorn[standard]==0.32.1
pydantic==2.10.6
pydantic-settings==2.5.0
python-multipart==0.0.7

# Database
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
alembic==1.14.1

# Energy Data APIs
entsoe-py==0.7.8

# Data Processing
lxml>=4.9.0
numpy==2.3.5
pandas==2.3.3
python-dateutil==2.8.2

# Async & Scheduling
httpx==0.28.1
aiohttp==3.11.12
tenacity==8.2.3

# Caching
cachetools==5.3.2

# Environment & Config
python-dotenv==1.0.1

# Production Server
gunicorn>=21.0.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2